test(openomni,docs): pr 561 review follow-ups
CodeRabbit follow-ups on merged #561 (items 2-4; item 1 is the
previous commit):

- docs/implementation-status.md: align every PendingInteraction row
  with the frozen/read-only + Wait-owned reality. The PendingInteraction
  row drops the stale "status transitions only" title and the
  resolve/markFollowUp + boot-expiry-sweep claims (write surface fully
  frozen behind the typed FrozenError, expiry is a read-time correlation
  filter, boot emits a frozen no-op receipt, table is in the archive
  manifest); the question-bridge row stops promising Owner-facing PI
  creation (new questions are Wait-owned); the connector row scopes
  report_result routing to frozen legacy rows; the #510 ledger row's D2b
  scope drops the already-frozen PendingInteraction half.
- test fixtures: the validated adapter-level legacy-row insertion is
  extracted into the one shared helper
  test/helpers/pending-interaction.ts and used from both the dispatch
  runtime fixtures' former copy and the ingress kernel-routing seeder.
- kernel-routing-waits: the #548 frozen-row regression pin now asserts
  status === "open" (read-only, untransitioned) instead of mere
  row existence.
