+++ packages/openomni/test/dispatch/runtime-pending-interaction-denial.test.ts
@@ -1,11 +1,8 @@
 import { beforeEach, describe, expect, test } from "bun:test";
 import { PendingInteractionStore, Storage } from "@openomni/session";
 import { DispatchRuntime } from "../../src/dispatch/runtime";
-import {
-  createWorkerRunFixture,
-  resetDispatchTestState,
-  seedPendingInteraction,
-} from "./runtime-test-fixtures";
+import { seedPendingInteraction } from "../helpers/pending-interaction";
+import { createWorkerRunFixture, resetDispatchTestState } from "./runtime-test-fixtures";
 
 describe("DispatchRuntime", () => {
   beforeEach(resetDispatchTestState);
